+++ CHANGES.md
@@ -4,6 +4,9 @@
 -   Updated documentation for setting up `Docker Edge` for Kubernetes.
 -   Modified error message text and omitted homepage articles if search error occurs.
 -   Changed chart icons on dataset page
+-   Made `elasticsearch` supports synonyms
+-   Upgraded `elasticsearch` to v6.3.0
+-   Enable organisation search
 -   Disabled scroll to zoom on dataset preview map until map is clicked on and added terria zoom controls to preview map
 -   Change appearance of mobile search box
 -   Improve filter button icon alignment